+++ frontend/app/players/[tag]/page.tsx
@@ -13,6 +13,7 @@ import BattleLogSoloRanked from "@/app/players/[tag]/_components/BattleLogSoloRa
 import BattleLogTrio from "@/app/players/[tag]/_components/BattleLogTrio";
 import { Telemetry } from "@/app/players/[tag]/_components/Telemetry.tsx";
 import styles from "./page.module.scss";
+import { formatBattleLog } from "@/app/_lib/formatBattleLog";
 
 type Player = {
   tag: string;
@@ -242,45 +243,3 @@ export default async function Page({
     </>
   );
 }
-
-export const formatBattleLog = (battleLogs: any[]) => {
-  const formattedBattleLogs: any[] = [];
-  battleLogs.forEach((battleLog: any) => {
-    if (battleLog.battle.type === "soloRanked") {
-      const existingBattle = formattedBattleLogs.find((b) => {
-        if (!b?.battle?.teams) {
-          return false;
-        }
-
-        const beforeBattleLogId = b.battle.teams
-          .flat()
-          .map((player: any) => player.tag)
-          .sort()
-          .join("-");
-        const currentBattleLogId = battleLog.battle.teams
-          .flat()
-          .map((player: any) => player.tag)
-          .sort()
-          .join("-");
-        return beforeBattleLogId === currentBattleLogId;
-      });
-      const roundData = {
-        battleTime: battleLog.battleTime,
-        result: battleLog.battle.result,
-        duration: battleLog.battle.duration,
-      };
-      if (existingBattle) {
-        existingBattle.rounds.push(roundData);
-        existingBattle.rounds.sort((a: any, b: any) =>
-          a.battleTime.localeCompare(b.battleTime),
-        );
-      } else {
-        battleLog.rounds = [roundData];
-        formattedBattleLogs.push(battleLog);
-      }
-    } else {
-      formattedBattleLogs.push(battleLog);
-    }
-  });
-  return formattedBattleLogs;
-};
